control_system: introduce osignal
The osignal is a new approach to creating conversions/transfer
functions which is easier to work with than the statically defined
ones with APIs. osignal converts a q31_t signal -1 to 1 to an
arbitrary output range with 64 bit signed range. This is useful for
converting an osignal so a pwm pulse or some value in ns, us, ms
etc, without needing to understand the q31_t format.

Signed-off-by: Bjarki Arge Andreasen <bjarki.andreasen@nordicsemi.no>